Supabase client file changed to js file, code set up in Home.jsx to view user profile data
