+++ frontend/src/supabaseClient.js
@@ -1,0 +1,4 @@
+import { createClient } from "@supabase/supabase-js";
+
+const supabase = createClient(import.meta.env.VITE_SUPABASE_URL, import.meta.env.VITE_SUPABASE_PUBLISHABLE_KEY);
+export default supabase
